Add fractionDigits function
The fractionDigits function determines the number of decimal places
of a coin price to display.

A price greater than or equal to 0.10 displays to two decimal places.
A price less than 0.10 will display the first non-zero digit
to the right of the decimal point, and the digit that follows it.